# tenant_platform/contracts/openapi/platform.yaml
openapi: 3.1.0
info:
  title: GenericAgent Tenant Platform Foundation API
  version: 0.1.0
  description: >
    Loopback foundation HTTP contract for session tasks. X-Platform-Admin-Token is
    required for this foundation-only API. No browser chat endpoint is exposed;
    the P0 user chat path remains BotTransportAdapter.
servers:
  - url: http://127.0.0.1:8080
    description: Local loopback foundation backend
security:
  - PlatformAdminToken: []
paths:
...
  /v1/tasks/{task_id}:
    get:
      security:
        - BearerToken: []
      operationId: getTask
      summary: Read task status and opaque result references
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task status envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/tasks/{task_id}/cancel:
    post:
      security:
        - BearerToken: []
      operationId: cancelTask
      summary: Request cooperative task cancellation
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Cancellation acceptance and current status
          content:
            application/json:
              schema:
                type: object
                required:
                  - accepted
                  - status
                properties:
                  accepted:
                    type: boolean
                  status:
                    type: string
                    enum: [queued, starting, running, succeeded, failed, cancelled, interrupted]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    PlatformAdminToken:
      type: apiKey
      in: header
      name: X-Platform-Admin-Token
      description: Required foundation-only loopback development token
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: platform-user-token
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid X-Platform-Admin-Token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Task or result not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current task state (e.g. result_ref mismatch)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - trace_id
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: SUBMIT_FAILED
        message:
          type: string
          description: Human-readable error message
        trace_id:
          type: string
          description: Correlation trace id for debugging
...
    TaskStatus:
      type: object
      required:
        - task_id
        - session_key
        - status
      properties:
        task_id:
          type: string
        session_key:
          type: string
        status:
          type: string
          enum: [queued, starting, running, succeeded, failed, cancelled, interrupted]
        snapshot_id:
          type: string
        snapshot_checksum:
          type: string
        result_ref:
          type: string
          description: Opaque committed result reference; never a host path
        result_digest:
          type: string
        terminal_error:
          type:
            - object
            - "null"
          properties:
            code:
              type: string
            user_message:
              type: string
            trace_id:
              type: string
...